Generate the slow clock in pong.v
Instead of taking 2 clocks as input lines in pong.v, take only the
fast 31.5MHz clock and divide it down to 10MHz in pong.v.

This because Caravel is not able to generate 2 clocks from its PLL
and so the divider would have be generated in the user project
wrapper module, which it seems cannot contain any always() blocks.