Quadruple the resolution to 32x32
This change doubles the screen resolution from 16 to 32 in both
dimensions. This is downsampled on the dot matrix display to 16x16
and on VGA upsampled.

This change also drops support for different size paddles, in favor
of just a single paddle size. Since the paddle array also doubled,
this is more space efficient as we're already starting to push the
die area capacity.