Added gate level simulation test
Also removed the paddle rom file in favor of a statically defined
localparam in the code. We no longer have multiple different paddle
sizes and so having it in the source file is not unwieldy.

Also added explicit reset code to various modules to facilitate the
gate level simulation process.