Freed up IO pins by externalizing 7-segment driver
The score is now exposed as a 4-bit binary number for conversion
off-chip by a 74HC4511, along with 2 common cathode pins for strobing
and blinking.

This frees up 2 IO pins, one of which has been allocated to the
difficulty dial, which was only using 3 of its 4 lines.

This leaves one IO line. Perhaps it can be used for a buzzer.